requirements.txt
streamlit>=1.22.0
requests>=2.28.1
Pillow>=9.2.0
python-dotenv>=1.0.0